+++ eclipse/plugins/net.sf.orcc.xdf.ui/src/net/sf/orcc/xdf/ui/diagram/XdfDiagramFeatureProvider.java
@@ -31,7 +31,9 @@ package net.sf.orcc.xdf.ui.diagram;
 import net.sf.orcc.df.Port;
 import net.sf.orcc.graph.Vertex;
 import net.sf.orcc.util.OrccLogger;
+import net.sf.orcc.xdf.ui.features.CopyFeature;
 import net.sf.orcc.xdf.ui.features.DropInstanceFromFileFeature;
+import net.sf.orcc.xdf.ui.features.PasteFeature;
 import net.sf.orcc.xdf.ui.features.UpdateDiagramFeature;
 import net.sf.orcc.xdf.ui.layout.AutoLayoutFeature;
 import net.sf.orcc.xdf.ui.patterns.ConnectionPattern;
@@ -43,13 +45,17 @@ import net.sf.orcc.xdf.ui.util.PropsUtil;
 import org.eclipse.core.resources.IFile;
 import org.eclipse.graphiti.dt.IDiagramTypeProvider;
 import org.eclipse.graphiti.features.IAddFeature;
+import org.eclipse.graphiti.features.ICopyFeature;
 import org.eclipse.graphiti.features.IFeature;
+import org.eclipse.graphiti.features.IPasteFeature;
 import org.eclipse.graphiti.features.IReconnectionFeature;
 import org.eclipse.graphiti.features.IRemoveFeature;
 import org.eclipse.graphiti.features.IUpdateFeature;
 import org.eclipse.graphiti.features.context.IAddContext;
 import org.eclipse.graphiti.features.context.IContext;
+import org.eclipse.graphiti.features.context.ICopyContext;
 import org.eclipse.graphiti.features.context.ICustomContext;
+import org.eclipse.graphiti.features.context.IPasteContext;
 import org.eclipse.graphiti.features.context.IPictogramElementContext;
 import org.eclipse.graphiti.features.context.IReconnectionContext;
 import org.eclipse.graphiti.features.context.IRemoveContext;
@@ -78,6 +84,8 @@ import de.cau.cs.kieler.kiml.options.LayoutOptions;
 public class XdfDiagramFeatureProvider extends
 		DefaultFeatureProviderWithPatterns {
 
+	private final PasteFeature pasteFeature;
+	private final CopyFeature copyFeature;
 	private final UpdateDiagramFeature updateFeature;
 	private final DropInstanceFromFileFeature dropInstanceFeature;
 	private final ICustomFeature[] layoutFeatures;
@@ -89,6 +97,8 @@ public class XdfDiagramFeatureProvider extends
 		addPattern(new OutputNetworkPortPattern());
 		addConnectionPattern(new ConnectionPattern());
 
+		copyFeature = new CopyFeature(this);
+		pasteFeature = new PasteFeature(this);
 		updateFeature = new UpdateDiagramFeature(this);
 		dropInstanceFeature = new DropInstanceFromFileFeature(this);
 
@@ -163,6 +173,16 @@ public class XdfDiagramFeatureProvider extends
 			return dropInstanceFeature;
 		}
 		return super.getAddFeature(context);
+	}
+
+	@Override
+	public ICopyFeature getCopyFeature(ICopyContext context) {
+		return copyFeature;
+	}
+
+	@Override
+	public IPasteFeature getPasteFeature(IPasteContext context) {
+		return pasteFeature;
 	}
 
 	/**